search on messages delete message from  me and all done

+++ v_ui_packages/v_chat_room_page/lib/src/message/page/message_page/message_provider.dart
@@ -81,7 +81,9 @@ class MessageProvider {
   }
 
   Future deleteMessageFromMe(VBaseMessage msg) async {
-    await _localMessage.deleteMessageByLocalId(msg);
+    if (msg.messageStatus.isServerConfirm) {
+      await _localMessage.deleteMessageByLocalId(msg);
+    }
     return _remoteMessage.deleteMessageFromMe(msg.roomId, msg.id);
   }
 }
